Moved common pizza model to :share:models
