+++ feature/pizza/build.gradle.kts
@@ -41,6 +41,7 @@ dependencies {
     implementation(project(":components:ui"))
     implementation(project(":components:theme"))
     implementation(project(":shared:network"))
+    implementation(project(":shared:model"))
     implementation(libs.coil.compose)
     implementation(platform(libs.androidx.compose.bom))
     implementation(libs.androidx.ui)
