+++ feature/pizza/src/main/java/ru/evgenykuzakov/pizza/data/mapper/DoughMapper.kt
@@ -1,9 +1,10 @@
 package ru.evgenykuzakov.pizza.data.mapper
 
-import ru.evgenykuzakov.cart.domain.model.DoughType
+import ru.evgenykuzakov.model.pizza.Dough
+import ru.evgenykuzakov.model.pizza.type.DoughType
 import ru.evgenykuzakov.pizza.data.model.DoughDto
 
-fun DoughDto.toDomain() = ru.evgenykuzakov.cart.domain.model.Dough(
+fun DoughDto.toDomain() = Dough(
     price = price,
-    type = ru.evgenykuzakov.cart.domain.model.DoughType.fromString(type)
+    type = DoughType.fromString(type)
 )
