+++ feature/pizza/src/main/java/ru/evgenykuzakov/pizza/data/mapper/SizeMapper.kt
@@ -1,9 +1,10 @@
 package ru.evgenykuzakov.pizza.data.mapper
 
-import ru.evgenykuzakov.cart.domain.model.SizeType
+import ru.evgenykuzakov.model.pizza.Size
+import ru.evgenykuzakov.model.pizza.type.SizeType
 import ru.evgenykuzakov.pizza.data.model.SizeDto
 
-fun SizeDto.toDomain() = ru.evgenykuzakov.cart.domain.model.Size(
+fun SizeDto.toDomain() = Size(
     price = price,
-    type = ru.evgenykuzakov.cart.domain.model.SizeType.fromString(type)
+    type = SizeType.fromString(type)
 )
