+++ feature/pizza/src/main/java/ru/evgenykuzakov/pizza/data/mapper/ToppingMapper.kt
@@ -1,10 +1,11 @@
 package ru.evgenykuzakov.pizza.data.mapper
 
-import ru.evgenykuzakov.cart.domain.model.IngredientType
+import ru.evgenykuzakov.model.pizza.Topping
+import ru.evgenykuzakov.model.pizza.type.IngredientType
 import ru.evgenykuzakov.pizza.data.model.ToppingDto
 
-fun ToppingDto.toDomain() = ru.evgenykuzakov.cart.domain.model.Topping(
+fun ToppingDto.toDomain() = Topping(
     img = img,
     price = price,
-    type = ru.evgenykuzakov.cart.domain.model.IngredientType.fromString(type)
+    type = IngredientType.fromString(type)
 )
