+++ settings.gradle.kts
@@ -31,3 +31,4 @@ include(":components:theme")
 include(":components:ui")
 include(":feature:pizza")
 include(":shared:cart")
+include(":shared:model")
