+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/data/mapper/IngredientMapper.kt
@@ -1,9 +1,11 @@
 package ru.evgenykuzakov.cart.data.mapper
 
 import ru.evgenykuzakov.cart.data.database.model.IngredientEntity
-import ru.evgenykuzakov.cart.domain.model.type.IngredientType
+import ru.evgenykuzakov.model.pizza.Ingredient
+import ru.evgenykuzakov.model.pizza.IngredientType
+import ru.evgenykuzakov.model.pizza.type.IngredientType
 
-fun IngredientEntity.toDomain() = ru.evgenykuzakov.cart.domain.model.Ingredient(
+fun IngredientEntity.toDomain() = Ingredient(
     img = img,
     price = price,
     type = IngredientType.fromString(type)
