+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/data/mapper/SizeMapper.kt
@@ -1,8 +1,8 @@
 package ru.evgenykuzakov.cart.data.mapper
 
 import ru.evgenykuzakov.cart.data.database.model.SizeEntity
-import ru.evgenykuzakov.cart.domain.model.Size
-import ru.evgenykuzakov.cart.domain.model.type.SizeType
+import ru.evgenykuzakov.model.pizza.Size
+import ru.evgenykuzakov.model.pizza.type.SizeType
 
 fun SizeEntity.toDomain() = Size(
     price = price,
