--- shared/cart/src/main/java/ru/evgenykuzakov/cart/domain/model/Dough.kt
@@ -1,8 +1,0 @@
-package ru.evgenykuzakov.cart.domain.model
-
-import ru.evgenykuzakov.cart.domain.model.type.DoughType
-
-data class Dough(
-    val price: Int,
-    val type: DoughType
-)
