+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/domain/use_case/AddToCartUseCase.kt
@@ -1,9 +1,9 @@
 package ru.evgenykuzakov.cart.domain.use_case
 
-import ru.evgenykuzakov.cart.domain.model.Pizza
+import ru.evgenykuzakov.model.pizza.Pizza
 import ru.evgenykuzakov.cart.domain.repository.CartRepository
 import javax.inject.Inject
 
 class AddToCartUseCase @Inject constructor(
     private val repository: CartRepository
-) : suspend (Pizza) -> Unit by repository::addToCard
+) : suspend (ru.evgenykuzakov.model.pizza.Pizza) -> Unit by repository::addToCard
