+++ shared/model/src/main/java/ru/evgenykuzakov/model/pizza/Dough.kt
@@ -1,0 +1,8 @@
+package ru.evgenykuzakov.model.pizza
+
+import ru.evgenykuzakov.model.pizza.type.DoughType
+
+data class Dough(
+    val price: Int,
+    val type: DoughType
+)
