+++ shared/model/src/main/java/ru/evgenykuzakov/model/pizza/Ingredient.kt
@@ -1,0 +1,9 @@
+package ru.evgenykuzakov.model.pizza
+
+import ru.evgenykuzakov.model.pizza.type.IngredientType
+
+data class Ingredient(
+    val img: String,
+    val price: Int,
+    val type: IngredientType
+)
