+++ shared/model/src/main/java/ru/evgenykuzakov/model/pizza/Pizza.kt
@@ -1,4 +1,4 @@
-package ru.evgenykuzakov.cart.domain.model
+package ru.evgenykuzakov.model.pizza
 
 data class Pizza(
     val id: String,
