+++ shared/model/src/main/java/ru/evgenykuzakov/model/pizza/Size.kt
@@ -1,0 +1,8 @@
+package ru.evgenykuzakov.model.pizza
+
+import ru.evgenykuzakov.model.pizza.type.SizeType
+
+data class Size(
+    val price: Int,
+    val type: SizeType
+)
