+++ shared/model/src/main/java/ru/evgenykuzakov/model/pizza/Topping.kt
@@ -1,0 +1,9 @@
+package ru.evgenykuzakov.model.pizza
+
+import ru.evgenykuzakov.model.pizza.type.IngredientType
+
+data class Topping(
+    val img: String,
+    val price: Int,
+    val type: IngredientType
+)
